requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
flake8==7.1.0
mccabe==0.7.0
mypy==1.11.1
mypy-extensions==1.0.0
numpy==2.0.1
pandas==2.2.2
psycopg2==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
